models/binance_adapter.py
# binance_adapter.py
from exchanges import Exchanges
from binance.client import Client

class BinanceAdapter(Exchanges):
    def __init__(self, api_key, secret_key):
        self.client = Client(api_key, secret_key)

    def buyAPI(self, amount, symbol):
        # Placeholder code for buyAPI for Binance
        order = self.client.create_order(
            symbol=symbol,
            side=Client.SIDE_BUY,
            type=Client.ORDER_TYPE_MARKET,
            quantity=amount
        )
        return order

    def getPriceAPI(self, symbol):
        # Implementation of getPriceAPI for Binance
        ticker = self.client.get_symbol_ticker(symbol=symbol)
        return float(ticker['price'])

    def sellAPI(self, amount, symbol):
        # Placeholder code for sellAPI for Binance
        order = self.client.create_order(
            symbol=symbol,
            side=Client.SIDE_SELL,
            type=Client.ORDER_TYPE_MARKET,
            quantity=amount
        )
        return order

    def execute_exchange(self, exchange):
        # Placeholder code for execute_exchange for Binance
        if exchange == 'buy':
            # Execute buy on Binance
            pass
        elif exchange == 'sell':
            # Execute sell on Binance
            pass
        else:
            # Handle invalid exchange type
            pass

    def set_authorization(self, api_key, secret_key):
        # Implementation of set_authorization for Binance
        self.client = Client(api_key, secret_key)
