requirements.txt
python-binance==0.7.11